+++ controller/subroutines/calculus.cpp
@@ -1,0 +1,9 @@
+
+using namespace std;
+
+namespace calculus
+{
+    intervalsCreator()
+    {
+    }
+}
